wd_regsrv.py - fix

+++ wd_regsrv.py
@@ -1,9 +1,9 @@
-r"""A utility to cleanup the WatchDox client installation and register with a new server
+'''A utility to cleanup the WatchDox client installation and register with a new server
 
 This utility will:
 - Delete the entire %AppData%\WatchDox folder
 - Set the one-time registry key with the server (After asking the user which server he wants to connect 
-"""
+'''
 
 import _winreg
 import os
@@ -11,20 +11,20 @@ import shutil
 
 servers = [
     ('qa', 'qa.watchdox.com'),
-    ('srv30', 'srv30.watchdox.com')
+    ('srv30', 'srv30.watchdox.com'),
+    ('production', 'watchdox.com')
 ]
 
 
 def getServerFromUser():
     "propmt the user list of servers, and get the server he wants to install"
-
     print "Select the server that you want to register\n"
     for i,srvtpl in enumerate(servers):
         short_name, server_name = srvtpl
-        print "{0:-2}. {1}({2})".format(i, server_name, short_name)
+        print "%2d. %s(%s)" % (i, server_name, short_name)
     print "{0:2}. Cancel".format( len(servers) )
 
-    print '\n\n'
+    print '\n'
     choice = raw_input("Your choise: ")
     try:
         choice = int(choice)
@@ -39,13 +39,15 @@ def getServerFromUser():
 
 
 def setRegistry():
-    regkey = _winreg.OpenKey( _winreg.HKEY_CURRENT_USER, r'Software\WatchDox\OfficePlugin', 0, _winreg.KEY_WRITE )
-    _winreg.SetValueEx( regkey, 'ServerURL_Hidden', 0, _winreg.REG_SZ, server_name )
+    regkey = _winreg.OpenKey( _winreg.HKEY_CURRENT_USER, 'Software\\WatchDox\\OfficePlugin', 0, _winreg.KEY_WRITE )
+    _winreg.SetValueEx( regkey, 'ServerURL_Hidden', 0, _winreg.REG_SZ, server_name )	
+    _winreg.SetValueEx( regkey, 'InstallationPath', 0, _winreg.REG_SZ, 'c:\\watchdox\\trunk\\WindowsPlugin\\Debug\\' )
 
 def cleanFiles():
     adata = os.environ['APPDATA']
-    wddata = adata + 'WatchDox'
-    shutil.rmtree( wddata )
+    wddata = adata + '\\WatchDox'
+    if os.path.isdir( wddata ):
+        shutil.rmtree( wddata )
         
 if __name__ == '__main__':
     short_name, server_name = getServerFromUser()
